// src/modules/provider/provider.interface.ts
export interface ICreateProviderProfilePayload {
  bio?: string;
  experienceYears?: number;
  serviceArea?: string;
  latitude?: number;
  longitude?: number;
}

export interface IUpdateProviderProfilePayload {
  bio?: string;
  experienceYears?: number;
  serviceArea?: string;
  latitude?: number;
  longitude?: number;
}